tools/index-viz/pyproject.toml
[project]
name = "index-viz"
version = "0.1.0"
description = "Visualize the gf ask vector index as an interactive 2D map"
requires-python = ">=3.12"
dependencies = [
    "numpy>=1.26",
    "umap-learn>=0.5",
    "plotly>=5.18",
    "click>=8.1",
    "rich>=13.0",
]

[project.scripts]
index-viz = "index_viz.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/index_viz"]
